PP-4439 Remove TransactionFlow
- Replaced transaction flow with regular methods annotated with @Transactional
- Methods annotated with @Transactional need to be public, otherwise they will
not get picked by GuicePersist
