+++ src/main/java/uk/gov/pay/connector/charge/service/ChargeExpiryService.java
@@ -2,7 +2,7 @@ package uk.gov.pay.connector.charge.service;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
-import com.google.inject.Provider;
+import com.google.inject.persist.Transactional;
 import org.apache.commons.lang3.tuple.Pair;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -12,13 +12,16 @@ import uk.gov.pay.connector.charge.dao.ChargeDao;
 import uk.gov.pay.connector.charge.exception.ChargeNotFoundRuntimeException;
 import uk.gov.pay.connector.charge.model.domain.ChargeEntity;
 import uk.gov.pay.connector.charge.model.domain.ChargeStatus;
-import uk.gov.pay.connector.charge.service.transaction.TransactionContext;
-import uk.gov.pay.connector.charge.service.transaction.TransactionFlow;
-import uk.gov.pay.connector.charge.service.transaction.TransactionalOperation;
 import uk.gov.pay.connector.chargeevent.dao.ChargeEventDao;
+import uk.gov.pay.connector.common.exception.ConflictRuntimeException;
+import uk.gov.pay.connector.common.exception.IllegalStateRuntimeException;
+import uk.gov.pay.connector.common.exception.OperationAlreadyInProgressRuntimeException;
 import uk.gov.pay.connector.gateway.PaymentProviders;
+import uk.gov.pay.connector.gateway.model.request.CancelGatewayRequest;
 import uk.gov.pay.connector.gateway.model.response.BaseCancelResponse;
 import uk.gov.pay.connector.gateway.model.response.GatewayResponse;
+import uk.gov.pay.connector.gatewayaccount.model.GatewayAccountEntity;
+import uk.gov.pay.connector.paymentprocessor.model.OperationType;
 
 import javax.inject.Inject;
 import java.time.ZonedDateTime;
@@ -29,16 +32,15 @@ import java.util.Map;
 import java.util.stream.Collectors;
 
 import static com.google.common.collect.Lists.newArrayList;
+import static uk.gov.pay.connector.charge.model.domain.ChargeStatus.AUTHORISATION_3DS_READY;
 import static uk.gov.pay.connector.charge.model.domain.ChargeStatus.AUTHORISATION_3DS_REQUIRED;
+import static uk.gov.pay.connector.charge.model.domain.ChargeStatus.AUTHORISATION_READY;
 import static uk.gov.pay.connector.charge.model.domain.ChargeStatus.AUTHORISATION_SUCCESS;
 import static uk.gov.pay.connector.charge.model.domain.ChargeStatus.AWAITING_CAPTURE_REQUEST;
 import static uk.gov.pay.connector.charge.model.domain.ChargeStatus.CREATED;
 import static uk.gov.pay.connector.charge.model.domain.ChargeStatus.ENTERING_CARD_DETAILS;
 import static uk.gov.pay.connector.charge.model.domain.ChargeStatus.EXPIRED;
 import static uk.gov.pay.connector.charge.model.domain.ChargeStatus.EXPIRE_CANCEL_FAILED;
-import static uk.gov.pay.connector.charge.service.CancelServiceFunctions.changeStatusTo;
-import static uk.gov.pay.connector.charge.service.CancelServiceFunctions.doGatewayCancel;
-import static uk.gov.pay.connector.charge.service.CancelServiceFunctions.prepareForTerminate;
 import static uk.gov.pay.connector.charge.service.StatusFlow.EXPIRE_FLOW;
 
 public class ChargeExpiryService {
@@ -48,18 +50,17 @@ public class ChargeExpiryService {
     private static final String EXPIRY_SUCCESS = "expiry-success";
     private static final String EXPIRY_FAILED = "expiry-failed";
 
-    public static final List<ChargeStatus> EXPIRABLE_REGULAR_STATUSES = ImmutableList.of(
+    static final List<ChargeStatus> EXPIRABLE_REGULAR_STATUSES = ImmutableList.of(
             CREATED,
             ENTERING_CARD_DETAILS,
             AUTHORISATION_3DS_REQUIRED,
             AUTHORISATION_SUCCESS);
 
-    public static final List<ChargeStatus> EXPIRABLE_AWAITING_CAPTURE_REQUEST_STATUS = ImmutableList.of(AWAITING_CAPTURE_REQUEST);
+    static final List<ChargeStatus> EXPIRABLE_AWAITING_CAPTURE_REQUEST_STATUS = ImmutableList.of(AWAITING_CAPTURE_REQUEST);
 
     private final ChargeDao chargeDao;
     private final ChargeEventDao chargeEventDao;
     private final PaymentProviders providers;
-    private final Provider<TransactionFlow> transactionFlowProvider;
     static final List<ChargeStatus> GATEWAY_CANCELLABLE_STATUSES = Arrays.asList(
             AUTHORISATION_SUCCESS,
             AWAITING_CAPTURE_REQUEST
@@ -70,16 +71,14 @@ public class ChargeExpiryService {
     public ChargeExpiryService(ChargeDao chargeDao,
                                ChargeEventDao chargeEventDao,
                                PaymentProviders providers,
-                               Provider<TransactionFlow> transactionFlowProvider,
                                ConnectorConfiguration config) {
         this.chargeDao = chargeDao;
         this.chargeEventDao = chargeEventDao;
         this.providers = providers;
-        this.transactionFlowProvider = transactionFlowProvider;
         this.chargeSweepConfig = config.getChargeSweepConfig();
     }
 
-    public Map<String, Integer> expire(List<ChargeEntity> charges) {
+    Map<String, Integer> expire(List<ChargeEntity> charges) {
         Map<Boolean, List<ChargeEntity>> chargesToProcessExpiry = charges
                 .stream()
                 .collect(Collectors.partitioningBy(chargeEntity ->
@@ -106,10 +105,7 @@ public class ChargeExpiryService {
 
     private int expireChargesWithCancellationNotRequired(List<ChargeEntity> nonAuthSuccessCharges) {
         List<ChargeEntity> processedEntities = nonAuthSuccessCharges
-                .stream().map(chargeEntity -> transactionFlowProvider.get()
-                        .executeNext(changeStatusTo(chargeDao, chargeEventDao, chargeEntity.getExternalId(), EXPIRED))
-                        .complete()
-                        .get(ChargeEntity.class))
+                .stream().map(chargeEntity -> changeStatusTo(chargeEntity.getExternalId(), EXPIRED))
                 .collect(Collectors.toList());
 
         return processedEntities.size();
@@ -119,34 +115,18 @@ public class ChargeExpiryService {
 
         final List<ChargeEntity> expireCancelled = newArrayList();
         final List<ChargeEntity> expireCancelFailed = newArrayList();
-        final List<ChargeEntity> unexpectedStatuses = newArrayList();
 
         gatewayAuthorizedCharges.forEach(chargeEntity -> {
-            ChargeEntity processedEntity = transactionFlowProvider.get()
-                    .executeNext(prepareForTerminate(chargeDao, chargeEventDao, chargeEntity.getExternalId(), EXPIRE_FLOW))
-                    .executeNext(doGatewayCancel(providers))
-                    .executeNext(finishExpireCancel())
-                    .complete().get(ChargeEntity.class);
+            ChargeEntity processedEntity = prepareForTermination(chargeEntity.getExternalId(), EXPIRE_FLOW);
+            GatewayResponse gatewayResponse = doGatewayCancel(processedEntity);
+            ChargeEntity expiredCharge = finishExpireCancel(processedEntity.getExternalId(), gatewayResponse);
 
-            if (processedEntity == null) {
-                //this shouldn't happen, but don't break the expiry job
-                logger.error("Transaction context did not return a processed ChargeEntity during expiry of charge - charge_external_id={}",
-                        chargeEntity.getExternalId());
-            } else {
-                if (EXPIRED.getValue().equals(processedEntity.getStatus())) {
-                    expireCancelled.add(processedEntity);
-                } else if (EXPIRE_CANCEL_FAILED.getValue().equals(processedEntity.getStatus())) {
-                    expireCancelFailed.add(processedEntity);
-                } else {
-                    unexpectedStatuses.add(processedEntity); //this shouldn't happen, but still don't break the expiry job
-                }
+            if (EXPIRED.getValue().equals(expiredCharge.getStatus())) {
+                expireCancelled.add(processedEntity);
+            } else if (EXPIRE_CANCEL_FAILED.getValue().equals(expiredCharge.getStatus())) {
+                expireCancelFailed.add(expiredCharge);
             }
         });
-
-        unexpectedStatuses.forEach(chargeEntity ->
-                logger.error("ChargeEntity returned with unexpected status during expiry - charge_external_id={}, status={}",
-                        chargeEntity.getExternalId(), chargeEntity.getStatus())
-        );
 
         return Pair.of(
                 expireCancelled.size(),
@@ -171,19 +151,16 @@ public class ChargeExpiryService {
         }).orElse(statusFlow.getFailureTerminalState());
     }
 
-    private TransactionalOperation<TransactionContext, ChargeEntity> finishExpireCancel() {
-        return context -> {
-            String externalId = context.get(ChargeEntity.class).getExternalId();
-            return chargeDao.findByExternalId(externalId).map(chargeEntity -> {
-                GatewayResponse gatewayResponse = context.get(GatewayResponse.class);
-                ChargeStatus status = determineTerminalState(chargeEntity, gatewayResponse, EXPIRE_FLOW);
-                logger.info("Charge status to update - charge_external_id={}, status={}, to_status={}",
-                        chargeEntity.getExternalId(), chargeEntity.getStatus(), status);
-                chargeEntity.setStatus(status);
-                chargeEventDao.persistChargeEventOf(chargeEntity);
-                return chargeEntity;
-            }).orElseThrow(() -> new ChargeNotFoundRuntimeException(externalId));
-        };
+    @Transactional
+    public ChargeEntity finishExpireCancel(String chargeExternalId, GatewayResponse gatewayResponse) {
+        return chargeDao.findByExternalId(chargeExternalId).map(chargeEntity -> {
+            ChargeStatus status = determineTerminalState(chargeEntity, gatewayResponse, EXPIRE_FLOW);
+            logger.info("Charge status to update - charge_external_id={}, status={}, to_status={}",
+                    chargeEntity.getExternalId(), chargeEntity.getStatus(), status);
+            chargeEntity.setStatus(status);
+            chargeEventDao.persistChargeEventOf(chargeEntity);
+            return chargeEntity;
+        }).orElseThrow(() -> new ChargeNotFoundRuntimeException(chargeExternalId));
     }
 
     private void logUnsuccessfulResponseReasons(ChargeEntity chargeEntity, GatewayResponse gatewayResponse) {
@@ -202,5 +179,70 @@ public class ChargeExpiryService {
         int chargeExpiryWindowSeconds = chargeSweepConfig.getAwaitingCaptureExpiryThreshold();
         logger.debug("Charge expiry window size for awaiting_delay_capture in seconds: [{}]", chargeExpiryWindowSeconds);
         return ZonedDateTime.now().minusSeconds(chargeExpiryWindowSeconds);
+    }
+
+    private static String getLegalStatusNames(List<ChargeStatus> legalStatuses) {
+        return legalStatuses.stream().map(ChargeStatus::toString).collect(Collectors.joining(", "));
+    }
+
+    private static boolean chargeIsInTerminatableStatus(StatusFlow statusFlow, ChargeStatus chargeStatus) {
+        return statusFlow.getTerminatableStatuses().contains(chargeStatus);
+    }
+
+    private GatewayResponse doGatewayCancel(ChargeEntity chargeEntity) {
+        return providers.byName(chargeEntity.getPaymentGatewayName())
+                .cancel(CancelGatewayRequest.valueOf(chargeEntity));
+    }
+
+    // Only methods marked as public will be picked up by GuicePersist
+    @Transactional
+    public ChargeEntity changeStatusTo(String chargeId, ChargeStatus targetStatus) {
+        return chargeDao.findByExternalId(chargeId)
+                .map(chargeEntity -> {
+                    logger.info("Charge status to update - charge_external_id={}, status={}, to_status={}",
+                            chargeEntity.getExternalId(), chargeEntity.getStatus(), targetStatus);
+                    chargeEntity.setStatus(targetStatus);
+                    chargeEventDao.persistChargeEventOf(chargeEntity);
+                    return chargeEntity;
+                })
+                .orElseThrow(() -> new ChargeNotFoundRuntimeException(chargeId));
+    }
+
+    @Transactional
+    public ChargeEntity prepareForTermination(String chargeId, StatusFlow statusFlow) {
+        return chargeDao.findByExternalId(chargeId).map(chargeEntity -> {
+            ChargeStatus newStatus = statusFlow.getLockState();
+            final ChargeStatus chargeStatus = ChargeStatus.fromString(chargeEntity.getStatus());
+            if (!chargeIsInTerminatableStatus(statusFlow, chargeStatus)) {
+                if (newStatus.equals(chargeStatus)) {
+                    throw new OperationAlreadyInProgressRuntimeException(statusFlow.getName(), chargeId);
+                } else if (Arrays.asList(AUTHORISATION_READY, AUTHORISATION_3DS_READY).contains(chargeStatus)) {
+                    throw new ConflictRuntimeException(chargeEntity.getExternalId());
+                }
+
+                logger.error("Charge is not in one of the legal states. charge_external_id={}, status={}, legal_states={}",
+                        chargeId, chargeEntity.getStatus(), getLegalStatusNames(statusFlow.getTerminatableStatuses()));
+
+                throw new IllegalStateRuntimeException(chargeId);
+            }
+            chargeEntity.setStatus(newStatus);
+
+            GatewayAccountEntity gatewayAccount = chargeEntity.getGatewayAccount();
+
+            logger.info("Card cancel request sent - charge_external_id={}, charge_status={}, account_id={}, transaction_id={}, amount={}, operation_type={}, provider={}, provider_type={}, locking_status={}",
+                    chargeEntity.getExternalId(),
+                    chargeStatus,
+                    gatewayAccount.getId(),
+                    chargeEntity.getGatewayTransactionId(),
+                    chargeEntity.getAmount(),
+                    OperationType.CANCELLATION.getValue(),
+                    gatewayAccount.getGatewayName(),
+                    gatewayAccount.getType(),
+                    newStatus);
+
+            chargeEventDao.persistChargeEventOf(chargeEntity);
+
+            return chargeEntity;
+        }).orElseThrow(() -> new ChargeNotFoundRuntimeException(chargeId));
     }
 }
