--- src/main/java/uk/gov/pay/connector/charge/service/transaction/TransactionContext.java
@@ -1,36 +1,0 @@
-package uk.gov.pay.connector.charge.service.transaction;
-
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
-
-/**
- * Context class to be used by {@link TransactionFlow}
- * backed by {@link java.util.concurrent.ConcurrentHashMap}
- *
- * Collects intermediate parameter values during a transaction.
- * <b>Only a maximum one object of a given type can exist during a transaction flow.</b>
- */
-public class TransactionContext {
-
-    Map<Class<?>, Object> params;
-
-    /**
-     * only to be constructed by TransactionFlow
-     */
-    TransactionContext() {
-        params = new ConcurrentHashMap<>();
-    }
-
-    /**
-     * only to be used by TransactionFlow
-     * @param param
-     * @param <T>
-     */
-    <T> void put(T param) {
-        params.put(param.getClass(), param);
-    }
-
-    public <T> T get(Class<T> paramClass) {
-        return (T) params.get(paramClass);
-    }
-}
